requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.3.1
Babel==2.9.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.1
celery==4.4.7
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.11.2
django-firebase-authentication==1.0.9
django-google-analytics-app==5.0.3
django-js-asset==1.2.2
django-nested-admin==3.3.3
django-phonenumber-field==5.0.0
django-sitemaps==1.1.1
djangorestframework==3.12.2
firebase-admin==4.5.1
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.46.3
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
ipaddr==2.2.0
kombu==4.6.11
lockfile==0.12.2
lxml==4.9.0
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.16
Pillow==9.1.1
progress==1.5
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pyparsing==2.4.6
python-dotenv==0.18.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
whitenoise==5.2.0
